requirements.txt
asgiref==3.6.0
Django==4.1.7
django-authority==0.14
django-extensions==3.2.1
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
Markdown==3.4.3
Pillow==9.4.0
PyJWT==1.7.1
pytz==2023.2
sqlparse==0.4.3
